update datum cyclus

+++ update_toernooi_cyclus.php
@@ -1,8 +1,15 @@
 <?php
-# 21feb20209          1.0.0       E. Hendrikx 
+# 21feb2020         1.0.0       E. Hendrikx 
 # Symptom:   		 Ongeldige datum geeft probleem met insert in toernooi_ontip
 # Problem:     	     None
 # Fix:               checkdate functie
+# Feature:           
+# Reference: 
+
+# 13mar2020          1.0.0       E. Hendrikx 
+# Symptom:   		 Delete werkt niet als gevolg van ongeldige datum check
+# Problem:     	     None
+# Fix:               ongeldige datum check alleen als er geen delete is aangevinkt
 # Feature:           
 # Reference: 
 
@@ -31,8 +38,9 @@ $toernooi        = $_POST['toernooi'];
 $delete          = $_POST['Delete'];
 $url             = 'beheer_cyclus_datums.php?toernooi='.$toernooi;
 
-
 include('mysqli.php');
+//include('action.php');
+
 
 $error = 0;
 $sql      = mysqli_query($con,"SELECT * from toernooi_datums_cyclus  where  Vereniging_id = ". $vereniging_id." and Toernooi ='".$toernooi."'  order by Datum" )     ; 
@@ -57,8 +65,6 @@ $error =1;
 
 
 $datum =   $_POST['datum_new'];
-echo $datum;
-
 $datum_maand = substr($datum ,5,2);
 $datum_jaar  = substr($datum ,0,4);
 $datum_dag   = substr($datum ,8,2);
@@ -67,7 +73,7 @@ $datum_dag   = substr($datum ,8,2);
 // checkdate ( int $month , int $day , int $year ) :
 $check = checkdate($datum_maand, $datum_dag, $datum_jaar);
 
-if ($check === false) { 
+if ($check === false and $delete  =='') { 
      /// reset naar vandaag
 	 $error =1;
 	 ?>
@@ -91,7 +97,7 @@ if (strlen($datum) != 10){
 <?php
 }
 
-if ($error == 0){
+if ($error == 0 and $delete  ==''){
 	
 $sql      = mysqli_query($con,"SELECT * from toernooi_datums_cyclus  where  Vereniging_id = ". $vereniging_id." and Toernooi ='".$toernooi."'  order by Datum" )     ; 
 	
@@ -126,12 +132,13 @@ if (isset ($_POST['datum_new'])   and $_POST['datum_new'] !='' and $_POST['datum
 
 }
 
-if ($delete !='')  {
+}  // end if
+if ($error == 0 and $delete !='')  {
   foreach ($delete as $deleteid)
    {
 	
-//	     echo "DELETE from npc_vereniging_teams  where Id =  ".$deleteid." <br>";
-		mysqli_query($con,"DELETE from toernooi_datums_cyclus  where Id =  ".$deleteid." ");
+    //echo "DELETE from npc_vereniging_teams  where Id =  ".$deleteid." <br>";
+	mysqli_query($con,"DELETE from toernooi_datums_cyclus  where Id =  ".$deleteid." ");
 	
 		   
    }// end foreach
@@ -139,7 +146,7 @@ if ($delete !='')  {
 
 
 redirect($url);	
-}  // end if
+
